display_linux and custom_video_xrandr refactoring
